project/sub/requirements.txt
redis
psycopg2
requests
python-dotenv
